Reintroduce build.assets/README.md
Commit f27af71d replaced build.assets/README.md with an artifact to be
bundled in the release tarball.  This restores the README.md and moves
the release tarball readme to release-tarball-README.md.

This makes a lot more sense, as GitHub surfaces the file named README.md
as the 'index' of each directory. And the prior README.md explained
nothing about the purpose of the build.assets directory.