Diversify robotest upgrade OSes.
When thinking about our upgrade scenarios, I realized that we'd added
scenarios specifically for certain customers, but we were running these
on ubuntu, instead of centos/rhel, which is what the customers use.

Updating "important versions in the field" to use redhat:7 was a no
brainer.  The other OSes are selected from a mix of what is available in
robotest:

  https://github.com/gravitational/robotest/blob/1f49957d353b78cc740ae2a40bf602ac58c8fd6e/assets/terraform/gce/os.tf